refactor(modularization): Phase 5 step 5-S5b — agent_runtime domain move + PromptTemplateOps (task #26)
Top-level rename of the ``aperag.agent_runtime`` package into the
canonical ``aperag.domains.agent_runtime`` home + FastAPI routes
merge into the domain ``api/routes.py``. Matches the Phase 3
``aperag.index`` → ``aperag.domains.indexing`` precedent (Step 2)
and the Phase 5 conversation sibling shape (5-S4*).

File moves (``git mv`` preserves history):

* ``aperag/agent_runtime/{runtime,services,storage,schemas}.py``
  → ``aperag/domains/agent_runtime/``
* ``aperag/views/agent_runtime.py`` → new
  ``aperag/domains/agent_runtime/api/routes.py``

G1 rewrites inside the moved domain files:

* ``runtime.py`` L32 ``aperag.db.models`` → sibling
  ``aperag.domains.agent_runtime.db.models`` (Phase 5 5-S2b has the
  enums at the canonical home).
* ``runtime.py`` L36 legacy ``aperag.service.prompt_template_service``
  hard-import → new ``PromptTemplateOps`` Protocol + ``_prompt_template_ops``
  DI slot + ``set_prompt_template_ops`` setter. ``aperag/app.py``
  wires a thin ``_PromptTemplateOpsAdapter`` that fans out to the
  legacy singleton (two async methods) + module-level
  ``build_agent_query_prompt`` helper (sync). Narrow surface matches
  the three runtime call sites.
* ``runtime.py`` L260 (pre-move) lazy import of
  ``aperag.service.chat_document_service`` → direct cross-domain
  ``aperag.domains.conversation.service.chat_document_service``
  (Phase 5 5-S4d moved it into conversation; msg=940bd884 simplification
  — no ``ChatDocumentOps`` seam needed, the Protocol is retired per
  ports.py note).
* ``runtime.py`` L52 / L576 ``view_models.AgentMessage`` → direct
  sibling ``aperag.domains.agent_runtime.schemas.AgentMessage`` (Phase 5
  5-S5a carved it). L545 ``view_models.Collection(...)`` → direct
  ``aperag.domains.knowledge_base.schemas.Collection`` (KB canonical
  home).
* ``services.py`` L33 split the ``aperag.db.models`` import into the
  sibling agent_runtime-domain module + the cross-domain
  ``aperag.domains.conversation.db.models.BotType``. ``view_models.BotConfig``
  → direct conversation-domain ``BotConfig``. L582 nested
  ``aperag.db.models.AgentTurn`` lazy import → sibling agent_runtime
  domain path.
* ``storage.py`` L19 schemas import → sibling agent_runtime domain path.
* ``schemas.py`` at the 5-S5a carved path now folds in the legacy
  envelope / enum classes (``AgentTurnEnvelope`` / ``CreateTurnRequest``
  / ``VisibleAgentState`` / etc.) so the domain package is
  self-contained. The carved ``AgentMessage`` and its
  ``_bind_view_models_reexports`` hook are preserved.

Ports:

* New ``PromptTemplateOps(Protocol)`` in
  ``aperag/domains/agent_runtime/ports.py`` — three methods, all
  ``@runtime_checkable`` + narrow.
* ``ChatDocumentOps`` is marked retired (Phase 6 removal candidate)
  but kept below ``__all__`` for in-flight branch references.

Legacy shims (every pre-migration import path keeps resolving):

* ``aperag/agent_runtime/__init__.py`` re-exports
  ``agent_runtime_manager`` + the seven envelope classes via the
  canonical domain path.
* ``aperag/agent_runtime/{runtime,services,storage,schemas}.py`` are
  recreated as thin submodule shims that re-export the full public
  surface from the domain home. Phase 6 cleanup deletes both the
  package and its submodule shims in one sweep.
* ``aperag/views/agent_runtime.py`` re-exports ``router`` + the
  ``runtime_manager`` alias (the latter preserves test-side
  ``monkeypatch.setattr(..., "runtime_manager", ...)`` shape while
  5-S8 migrates tests off the legacy path).

``aperag/app.py``:

* Single ``agent_runtime_router`` include sourced directly from the
  domain route module (replaces the former
  ``aperag.views.agent_runtime`` import).
* New ``_PromptTemplateOpsAdapter`` class + ``_ar_set_prompt_template_ops``
  wire-up at module-scope startup. 5-S8 CRITICAL_WIRINGS will pick
  up ``(agent_runtime_runtime, "_prompt_template_ops")``.

Verify:

* ``import aperag.app`` OK;
  ``aperag.domains.agent_runtime.runtime._prompt_template_ops`` is
  non-None; legacy ``aperag.agent_runtime.runtime.agent_runtime_manager``
  identity-matches the domain singleton.
* ``pytest tests/unit_test/`` 673 passed / 29 skipped / 1 deselected
  (the performance flake stays pre-existing). Two previously-skipped
  tests in ``test_agent_runtime_v3.py`` flipped to passing after
  their patch targets were rewritten to the canonical domain paths.
* ``uvx ruff check`` + ``uvx ruff format --check`` clean.
* ``scripts/export_openapi.py --check`` exit 0 — OpenAPI byte-stable
  against the pre-move snapshot. Agent-runtime SSE + JSON routes
  preserve shape.
* G1 on ``aperag/domains/agent_runtime/**``: zero ``aperag.db.models``
  / ``aperag.service.*`` / ``aperag.schema.view_models`` imports.
  G16 green (no ``User`` ORM imports anywhere in the new domain).
