+++ tests/unit_test/agent_runtime/test_agent_runtime_v3.py
@@ -4,8 +4,8 @@ from types import SimpleNamespace
 import pytest
 from sqlalchemy.exc import IntegrityError
 
-import aperag.agent_runtime.services as agent_runtime_services
-import aperag.agent_runtime.storage as agent_runtime_storage
+import aperag.domains.agent_runtime.services as agent_runtime_services
+import aperag.domains.agent_runtime.storage as agent_runtime_storage
 from aperag.agent_runtime.schemas import (
     AgentArtifactEnvelope,
     AgentTimelineEventEnvelope,
@@ -15,7 +15,7 @@ from aperag.agent_runtime.schemas import (
 )
 from aperag.agent_runtime.services import EventService, HistoryWriter, TurnService
 from aperag.db.models import AgentEventActor, AgentTurnStatus
-from aperag.views import agent_runtime as agent_runtime_view
+from aperag.domains.agent_runtime.api import routes as agent_runtime_view
 
 
 def _now():
